refactor(auth): refactor enums and dtos
