+++ auth-service/src/main/java/com/artsem/api/authservice/util/KeycloakGroup.java
@@ -9,5 +9,6 @@ public enum KeycloakGroup {
     DRIVER("DRIVER"),
     PASSENGER("PASSENGER"),
     MANAGER("MANAGER");
+
     private final String group;
 }
